feat(font): update font settings in Next.js library
The existing font settings for the Next.js library have been updated. Specifically, the weight of the Noto_Sans_JP font has been adjusted from '700' to '500'. Additionally, the Noto_Emoji font has been added with specified settings including a subset of 'emoji', a style of 'normal', a weight of '700', and a variable of '--font-noto-emoji'.