chore(ci): disable 'Checkout repository' and 'Build' steps in ci.yml workflow
The 'Checkout repository' and 'Build' steps in the ci.yml GitHub workflow file have been commented out. This has been done to prevent these steps from executing within the workflow, possibly for development, testing, or troubleshooting purposes. These changes are temporary and the steps are expected to be restored later.