fix(storybook): add new colors file and update Storybook scripts in package.json
A new colors file was created in the Storybook app with a 'Colors' title but remaining content to be added. Also, the 'storybook:move' script in the web app's package.json was adjusted to move the 'storybook-static' directory to a 'web' sub-directory. Three new scripts were added in the Storybook app's package.json to aid in development, building, and moving of Storybook app.