feat(post): add config-action component to posts
A ConfigAction component was added to the posts section to provide additional options such as viewing metadata, editing the page or reporting a content issue on GitHub, as well as viewing the source on GitHub. The layout was also updated to better accommodate the new component, adjusting the gap in between elements.