Staging deploy (#621)
* fix(deps): update dependency react-resizable-panels to v2.0.9 (#598)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* fix(deps): update dependency lucide-react to v0.331.0 (#609)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* fix(deps): update dependency @tanstack/react-table to v8.12.0 (#608)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* fix(deps): update turbo monorepo to v1.12.4 (#606)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* fix(deps): update dependency husky to v9.0.11 (#605)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* chore(deps): update dependency vite to v5.1.3 (#601)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* fix(deps): update dependency @vercel/analytics to v1.1.4 (#604)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* chore(deps): update node.js to v20.11.1 (#602)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* chore(deps): update storybook monorepo to v7.6.16 (#603)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* chore(public): add images

* chore(search): add new search item in the config

The searchItems config file was updated. The new entry represents a post about the replacement of an arm rest on a Sayl Chair. The new item includes its own link, emoji, and it falls under the 'Life' category.

* chore(content): sync submodule

* fix(deps): update dependency @vercel/analytics to v1.2.0 (#613)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* chore(deps): update dependency eslint-plugin-storybook to v0.8.0 (#612)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* chore(public): add post images

* chore(ui): refactor post navigation and link handling functions

Added comments describing the purpose and parameters for the functions 'getPager' in prev-and-next-pager/index.tsx, 'handleCopyLink' in action-controller/index.tsx, and 'fisherYatesShuffle' in the utils module. Moreover, the 'getPager' function was refactored to only consider published posts with an _id, without defaulting to null.

* chore(search): add v1.5.0 release notes link to searchItems configuration

Add another item to the static configuration list in searchItems.ts file which represents v1.5.0 release notes. The new item includes a title, href to the URL, an emoji representing tech, and a category indicating it's tech-related with a tag 'Release'.

* chore(content): sync submodule

* chore(pnpm): i

* chore(deps): update dependency @types/react to v18.2.56 (#616)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* chore(pnpm): i

* chore(ui): update font styling for prose and article titles

Added the 'palt' feature setting to typography in prose, enhancing the visual appeal and readability. This is also synced on the article title layout for consistency across the application.

* chore(format): reorder css classes in article title layout

The CSS classes in the article title layout were reordered to maintain consistency. The 'palt' class that adjusts the font style was moved to the beginning to enhance the visual appeal and readability of the application.

* chore(constant): add new tag

* chore(content): sync submodule

* chore(search): add a new tech post about integrating typeScript with play framework

The update includes a new tech post in the searchItems.ts file. The post discusses how to integrate TypeScript with the Play Framework and is categorized under 'Tech' with the tags 'Play Framework' and 'TypeScript'.

---------

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>