chore(ui): adjust margins for tabs-content in mdx components
The TabsContent component within global MDX components has been modified in order to fine-tune its margins. This commit introduces two changes. Firstly, the top margin of the first child is set to 0, and secondly, the bottom margin of the last child is set to 0. This ensures better alignment and spacing within the MDX UI.