feat(ui): add steps and step components to posts feature
A new Steps component and its child Step component have been added to the posts feature of the application. They are not only created in the posts' steps/index file but also imported into the posts index and the global MDX components file. In addition, the associated CSS for step counters has been added to the global stylesheet.