refactor(prettier): rearrange imports and tailwind classes
The code has been updated to enhance readability by rearranging the order of import statements. This has mostly involved moving type imports to the top, followed by library and module imports, and finally component imports. This change affects numerous files throughout the project and is meant to improve the organization and comprehension of the code.