refactor(tooling): update package.json for tailwind, eslint and prettier
Updated the package.json for tailwind to change from main to exports and added an ignore path for prettier checks in the scripts of tailwind, eslint and prettier. This will address the process of import resolution while respecting the project's unique path structure. Also, it allows to ignore files or folders during formatting according to the local .gitignore file.