refactor(ui): relocate MDX-Components and Code to Post UI
MDX-Components, formerly in the global UI, have been moved to the post UI directory for better organization. The 'Code' component has been renamed to 'CodeBlock' and has been refactored as 'figure' to the 'CodeBlock.' 'Step' and 'Steps' components are given additional props handling, and export adjustments are made according to the folder restructuring. This commit organizes the structure of the code according to its functionalities.