refactor(env): update environment variables configuration
Refactored the configuration of environment variables in `env.ts`. Import and extension of `authEnv` were added and `VERCEL_ENV` and `PORT` were removed. Also, default value for `CI` was changed to false and `DATABASE_URL` was added to the server configuration. Changes aimed at streamlining and updating the application environment setup.