chore(npm): update @playwright/test version from 1.43.1 to 1.44.0
This commit updates the version of the @playwright/test module that we're using in the package.json file. All references to the old version in the pnpm-lock.yaml file have also been updated to reflect this change. This ensures that our tests are running with the latest version of @playwright/test.