pref(api): add 'date-fns' package and use it in contact router
A 'date-fns' package has been added to package.json and lock file. It was subsequently used in the contact router for implementing a feature that adds a timestamp to each new contact input. The timestamp records the current date and time according to the 'yyyy/MM/dd HH:mm:ss' format.