pref(provider): add Google reCAPTCHA v3 integration
The Google reCAPTCHA v3 library has been integrated into the application to enhance security. The reCAPTCHA provider is wrapped around the application content and the invisible reCAPTCHA badge is styled. The package has also been added to the package.json file, and pnpm-lock updated accordingly. Also, the token field for reCAPTCHA validation is added to the contact form validators.