chore(api): replace console.error with throw for error handling
The error handling in various services has been updated to throw new Errors instead of just logging them to the console. This has been done in the line.ts, mail.ts, and the google/sheets.ts services. Additionally, some conditions have been added in the contact-form index.tsx to handle a scenario where the recaptcha token is not available or the mutation operation fails, displaying appriorate failure messages to the user.